dev startup

+++ update.sh
@@ -11,11 +11,13 @@ if ! [ -x "$(command -v docker-compose)" ]; then
 else
     DOCKER_COMPOSE_PATH="$(command -v docker-compose)"
 fi
-
-git reset --hard
-# pull the latest changes from the git repo and if there are any changes, restart the container (if there are no change git pull returns "Already up to date.")
-if [[ "$(git pull)" != "Already up to date." ]]; then
-    (cd "$(dirname "$0")/run" && $DOCKER_COMPOSE_PATH down && $DOCKER_COMPOSE_PATH build && $DOCKER_COMPOSE_PATH up -d)
+# if --dev is passed do not reset and pull the latest changes
+if [ "$1" != "--dev" ]; then
+    git reset --hard
+    # pull the latest changes from the git repo and if there are any changes, restart the container (if there are no change git pull returns "Already up to date.")
+    if [[ "$(git pull)" != "Already up to date." ]]; then
+        (cd "$(dirname "$0")/run" && $DOCKER_COMPOSE_PATH down && $DOCKER_COMPOSE_PATH build && $DOCKER_COMPOSE_PATH up -d)
+    fi
 fi
 
 # Check if there are any changes in the modules
